Group components in folders for cleaner code
