+++ src/Components/Charts/CoreVitalChart.jsx
@@ -9,12 +9,20 @@ import {
 } from "recharts";
 import { useState } from "react";
 import {
-  metricDescriptions,
   benchmarkLines,
   colors,
-} from "../config/chartConfig";
+  metricDescriptions,
+} from "../../config/chartConfig";
 
-export default function MetricChartWithToggles({ title, lines, data }) {
+export default function CoreVitalChart({
+  title,
+  description,
+  quality,
+  lines,
+  data,
+  yDomain,
+  unit,
+}) {
   const [visibleLines, setVisibleLines] = useState(() =>
     lines.map((line) => line.key)
   );
@@ -26,10 +34,13 @@ export default function MetricChartWithToggles({ title, lines, data }) {
   };
 
   return (
-    <div className="my-6 text-sm">
-      <h3 className="font-semibold text-gray-800 underline uppercase hidden">
+    <div className="my-8 text-sm">
+      <h3 className="font-semibold text-gray-800 uppercase underline mb-2">
         {title}
       </h3>
+      <p className="text-sm text-gray-700 rounded-lg mb-6 leading-relaxed italic">
+        {description} <br /> <span className="font-semibold">{quality}</span>
+      </p>
 
       {/* Toggle Checkboxes */}
       <div className="flex flex-wrap gap-4 mb-4">
@@ -66,57 +77,20 @@ export default function MetricChartWithToggles({ title, lines, data }) {
         ))}
       </div>
 
-      {/* Group Toggle Buttons */}
-      <div className="flex gap-4 mb-4 ml-1">
-        <button
-          onClick={() => {
-            const mobileKeys = lines
-              .filter((l) => l.device === "mobile")
-              .map((l) => l.key);
-            const allVisible = mobileKeys.every((k) =>
-              visibleLines.includes(k)
-            );
-            setVisibleLines((prev) =>
-              allVisible
-                ? prev.filter((k) => !mobileKeys.includes(k))
-                : [...new Set([...prev, ...mobileKeys])]
-            );
-          }}
-          className="px-3 py-1 bg-orange-500 text-white text-xs rounded hover:bg-orange-600 transition"
-        >
-          {lines.every(
-            (l) => l.device !== "mobile" || visibleLines.includes(l.key)
-          )
-            ? "Hide Mobile"
-            : "Show Mobile"}
-        </button>
-
-        <button
-          onClick={() => {
-            const desktopKeys = lines
-              .filter((l) => l.device === "desktop")
-              .map((l) => l.key);
-            const allVisible = desktopKeys.every((k) =>
-              visibleLines.includes(k)
-            );
-            setVisibleLines((prev) =>
-              allVisible
-                ? prev.filter((k) => !desktopKeys.includes(k))
-                : [...new Set([...prev, ...desktopKeys])]
-            );
-          }}
-          className="px-3 py-1 bg-green-600 text-white text-xs rounded hover:bg-green-700 transition"
-        >
-          {lines.every(
-            (l) => l.device !== "desktop" || visibleLines.includes(l.key)
-          )
-            ? "Hide Desktop"
-            : "Show Desktop"}
-        </button>
+      {/* Custom Legend */}
+      <div className="flex items-center gap-4 mb-3 ml-1">
+        <div className="flex items-center gap-2">
+          <span className="w-3 h-3 rounded-full bg-[#fb923c] inline-block" />
+          <span className="text-sm text-orange-400 font-medium">Mobile</span>
+        </div>
+        <div className="flex items-center gap-2">
+          <span className="w-3 h-3 rounded-full bg-[#22c55e] inline-block" />
+          <span className="text-sm text-green-500 font-medium">Desktop</span>
+        </div>
       </div>
 
       {/* Chart */}
-      <ResponsiveContainer width="100%" height={300}>
+      <ResponsiveContainer width="100%" height={260}>
         <LineChart data={data}>
           <XAxis
             dataKey="date"
@@ -129,16 +103,16 @@ export default function MetricChartWithToggles({ title, lines, data }) {
                 : ""
             }
           />
-          <YAxis />
+          <YAxis domain={yDomain} />
           <Tooltip
             formatter={(value, name) => [
-              value,
-              metricDescriptions[name] ||
-                lines.find((l) => l.key === name)?.label ||
-                name,
+              unit === "s"
+                ? `${value.toFixed(2)}s`
+                : unit === "ms"
+                ? `${value}ms`
+                : value,
+              metricDescriptions[name] || name,
             ]}
-            labelClassName="text-xs"
-            contentStyle={{ fontSize: "0.8rem" }}
           />
 
           {lines.map(
@@ -155,6 +129,7 @@ export default function MetricChartWithToggles({ title, lines, data }) {
                 />
               )
           )}
+
           {lines.map(
             ({ key }) =>
               visibleLines.includes(key) &&
@@ -162,11 +137,16 @@ export default function MetricChartWithToggles({ title, lines, data }) {
                 <ReferenceLine
                   key={`ref-${key}`}
                   y={benchmarkLines[key]}
-                  stroke="gray"
+                  stroke="red"
                   strokeDasharray="4"
                   label={{
                     position: "right",
-                    value: `Benchmark: ${benchmarkLines[key]}`,
+                    value:
+                      unit === "s"
+                        ? `Benchmark: ${benchmarkLines[key]}s`
+                        : unit === "ms"
+                        ? `Benchmark: ${benchmarkLines[key]}ms`
+                        : `Benchmark: ${benchmarkLines[key]}`,
                     fill: "#dc2626",
                     fontSize: 10,
                   }}
