+++ src/Components/Headers/SecondaryHeader.jsx
@@ -3,8 +3,8 @@ import { motion } from "framer-motion";
 import { BarChart3, Lightbulb, FolderOpen } from "lucide-react";
 import clsx from "clsx";
 
-import { FavouriteBtn } from "./ReportButtons";
-import useBreakpoint from "../hooks/useBreakpoint";
+import { FavouriteBtn } from "../Buttons/ReportButtons";
+import useBreakpoint from "../../hooks/useBreakpoint";
 
 export default function SecondaryHeader() {
   const navigate = useNavigate();
