+++ src/Components/ResultsBlocks/CurrResultsBlock.jsx
