+++ src/Components/Toggles/DeviceToggle.jsx
